// backend/src/common/enums/document-type.enum.ts
export enum DocumentType {
  PAN = "PAN",
  AADHAAR = "AADHAAR",
  PHOTO = "PHOTO",
  BANK_STATEMENT = "BANK_STATEMENT",
  ITR = "ITR",
  PROPERTY_DOCUMENT = "PROPERTY_DOCUMENT",
  INCOME_PROOF = "INCOME_PROOF",
  CUSTOMER_CONSENT = "CUSTOMER_CONSENT",
  UDYAM_CERTIFICATE = "UDYAM_CERTIFICATE",
  BUSINESS_LICENSE = "BUSINESS_LICENSE",
  OTHER = "OTHER",
}