// backend/src/common/enums/role.enum.ts
export enum RoleCode {
  ADMIN = 'ADMIN',
  RM = 'RM',
  BM = 'BM',
  CM = 'CM',
  CREDIT = 'CREDIT',
  CREDIT_MAKER = 'CREDIT_MAKER',
  CREDIT_CHECKER = 'CREDIT_CHECKER',
  LEGAL = 'LEGAL',
  VALUATION = 'VALUATION',
  SANCTION = 'SANCTION',
  AGREEMENT = 'AGREEMENT',
  OPS_MAKER = 'OPS_MAKER',
  OPS_HEAD = 'OPS_HEAD',
  OPS_CHECKER = 'OPS_CHECKER',
  DISBURSEMENT = 'DISBURSEMENT',
  FEE_OPS = 'FEE_OPS',
  PAYMENT_MAKER = 'PAYMENT_MAKER',
  PAYMENT_CHECKER = 'PAYMENT_CHECKER',
  LMS = 'LMS',
  COLLECTION = 'COLLECTION'
}