// backend/src/common/enums/workflow-action.enum.ts
export enum WorkflowAction {
  SAVE_DRAFT = 'SAVE_DRAFT',
  SUBMIT = 'SUBMIT',
  SUBMIT_TO_BM = 'SUBMIT_TO_BM',
  SEND_BACK = 'SEND_BACK',
  APPROVE = 'APPROVE',
  REJECT = 'REJECT'
}
export enum WorkflowLogAction {
  LEAD_CREATED = 'LEAD_CREATED',
  LEAD_SUBMITTED = 'LEAD_SUBMITTED',

  CUSTOMER_VISIT_DONE = 'CUSTOMER_VISIT_DONE',
  GEO_VERIFICATION_DONE = 'GEO_VERIFICATION_DONE',
  PROPERTY_VISIT_DONE = 'PROPERTY_VISIT_DONE',

  DOCUMENTS_UPLOADED = 'DOCUMENTS_UPLOADED',
  SUBMITTED_TO_BM = 'SUBMITTED_TO_BM',
}